fix: header component with loginstatus
